requirements.txt
mss>=9.0.1
Pillow>=10.2.0
rapidfuzz>=3.9.0
keyboard>=0.13.5
requests>=2.32.3
beautifulsoup4>=4.12.3
# OCR backend wrapper (requires native Tesseract executable installed)
pytesseract>=0.3.13
